[level 2] Title: k진수에서 소수 개수 구하기, Time: 13.95 ms, Memory: 72 MB -BaekjoonHub

+++ 프로그래머스/2/92335. k진수에서 소수 개수 구하기/k진수에서 소수 개수 구하기.java
@@ -12,7 +12,7 @@ class Solution {
     }
     
     public boolean isPrime(long n) {
-	for (int i = 2; i<=(int)Math.sqrt(n); i++) {
+	for (long i= 2; i<=Math.sqrt(n); i++) {
       if (n % i == 0) 
           return false;
 	    }
